+++ tests/integration/orchestrator/scenarios/fake_rl_job.go
@@ -1,4 +1,4 @@
-package acceleratororchestrator
+package scenarios
 
 import (
 	"context"
