+++ tests/integration/orchestrator/scenarios/pod_factory.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package acceleratororchestrator
+package scenarios
 
 import (
 	"sync"
